Weitere Anpassungen an Systemd-Services

+++ solarswarm_setup/system services/docker_leader.bash
@@ -24,48 +24,59 @@ REMOVE_DUPLICATE_HOSTNAMES=true
 # when bravo (the host) rejoined, it was given a new ID and the status 'Ready'
 # commands using the then ambiguous hostname bravo failed
 
-# function to check for and remove duplicate hostnames as hostnames are not unique in a swarm
 check_duplicate_hostnames() {
-    # optionally pass one hostname as argument to filter, i.e.: 'check_duplicate_hostnames alfa'
-    hostnames=$(sudo docker node ls --filter "name=$1" --format "{{.Hostname}}")
-    for name in $(cat $SW_SETUP/ssh_identities/names); do
-        occurances=$(echo $hostnames | grep -E "$name " | wc -w)
-        # hosts_ready=0; hosts_down=0; hosts_unknown=0; hosts_disconnected=0
-        if [ $occurances -le 1 ]; then continue; fi
-        if [ $occurances -gt 1 ]; then
-            statuses=$(sudo docker node ls --filter "name=$name" --format "{{.ID}} {{.Status}}")
-            # get number of IDs using the same hostname per status
-            # output is assigned as words in a single line
-            # grep returns a selection of lines which contain a certain word
-            # therefore replace spaces with newline characters, then use grep
-            # s to replace, g for global, '/ /\n/' to replace ' ' with '\n'
-            hosts_ready=$(echo $statuses | sed "s/ /\n/g" | grep "Ready" | wc -w);
-            hosts_down=$(echo $statuses | sed "s/ /\n/g" | grep "Down" | wc -w);
-            hosts_unknown=$(echo $statuses | sed "s/ /\n/g" | grep "Unknown" | wc -w);
-            hosts_disconnected=$(echo $statuses | sed "s/ /\n/g" | grep "Disconnected" | wc -w);
-            echo "[docker_leader] Warning: Found $occurances nodes using hostname $name" >>$LOG_OUT
-            echo "    Ready       : $hosts_ready" >>$LOG_OUT
-            echo "    Down        : $hosts_down" >>$LOG_OUT
-            echo "    Unknown     : $hosts_unknown" >>$LOG_OUT
-            echo "    Disconnected: $hosts_disconnected" >>$LOG_OUT
-            if [ $REMOVE_DUPLICATE_HOSTNAMES == true ] && [ $hosts_ready -eq 1 ]; then
-                sudo docker node ls --filter "name=$name" --format "{{.ID}} {{.Status}}" | \
-                while read id status; do
-                    # note that it might take a few seconds for the leader to consider a host 'Down' after they left
-                    # "Error response from daemon: [...] is not down and can't be removed"
-                    
-                    # if [ ! $status == "Ready" ]; then # unsure if 'Unknown' or 'Disconnected' count
-                    if [ $status == "Down" ]; then
-                        if sudo docker node rm -f $id &>/dev/null; then
-                            echo "[docker_leader] Removed node $id with duplicate hostname $name and status $status"
-                        fi
-                    fi
-                done
-            elif [ $hosts_ready -gt 1 ]; then
-                echo "[docker_leader] Error: Found $hosts_ready 'Ready' nodes using hostname $name. This means that multiple hosts were given the same hostname. If this error persists, please remove all but one of these nodes and set them up again with another name that is not already in use." >>$LOG_OUT
+    # check for and remove duplicats of hostname given as parameter
+    # return number of remaining duplicates or -1 if something goes wrong
+    if [ -z $1 ]; then
+        echo "[docker_leader] Check for duplicate hostnames failed: No hostname was provided" >>$LOG_OUT
+        return -1
+    fi
+
+    occurances=$(sudo docker node ls --filter "name=$1" --format "{{.Hostname}}" | wc -w)
+    if [ -z $occurances ]; then # docker node ls failed
+        return -1
+    elif [ $occurances -le 1 ]; then # no duplicates
+        return 0
+    fi
+
+    # at least one duplicate
+    nodes_ready=0; nodes_down=0; nodes_unknown=0; nodes_disconnected=0
+    if [ $occurances -gt 1 ]; then
+        statuses=$(sudo docker node ls --filter "name=$1" --format "{{.ID}}:{{.Status}}:{{.ManagerStatus}}")
+        # get number of IDs using the same hostname per status
+        for status in $statuses; do
+            read node_id status m_status <<< $(echo $status | sed 's/:/ /g')
+            if [ -z $node_id ] || [ -z $status ]; then continue; fi # skip empty
+            if [ $status == "Ready" ]; then
+                ((nodes_ready++))
+            elif [ $status == "Down" ]; then
+                ((nodes_down++))
+            elif [ $status == "Unknown" ]; then
+                ((nodes_unknown++))
+            elif [ $status == "Disconnected" ]; then
+                ((nodes_disconnected++))
             fi
+        done
+        echo "[docker_leader] Warning: Found $occurances nodes using hostname $name" >>$LOG_OUT
+        echo "    Ready       : $nodes_ready" >>$LOG_OUT
+        echo "    Down        : $nodes_down" >>$LOG_OUT
+        echo "    Unknown     : $nodes_unknown" >>$LOG_OUT
+        echo "    Disconnected: $nodes_disconnected" >>$LOG_OUT
+        if [ $REMOVE_DUPLICATE_HOSTNAMES == true ] && [ $nodes_ready -le 1 ]; then
+            # case: one Ready -> remove all others
+            # case: none Ready -> remove all
+            if [ $nodes_ready -le 1 ]; then
+                for status in $statuses; do
+                    read node_id status m_status <<< $(echo $status | sed 's/:/ /g')
+                    if [ $status == "Ready" ]; continue; fi # skip Ready
+                    if [ $m_status == "Reachable" ]; then sudo docker node demote $id; fi # demote if manager
+                    sudo docker node rm $id -f
+                done
+            fi
+        elif [ $nodes_ready -gt 1 ]; then
+            echo "[docker_leader] Error: Found $nodes_ready 'Ready' nodes using hostname $name. This means that multiple hosts were given the same hostname. If this error persists, please remove all but one of these nodes and set them up again with another name that is not already in use." >>$LOG_OUT
         fi
-    done
+    fi
 }
 
 remove_host_from_manager_list() {
@@ -248,63 +259,54 @@ while [ 0 ]; do
             fi
         fi
     done
-    
-    #####
-    # TODO: remove echo, sleep, and continue below and test administering managers 
-    # echo "[docker_leader] Debug: Loop cycle done (skipped managers) - sleeping for $LEADER_CHECK_DELAY seconds" >>$LOG_OUT
-    # sleep $LEADER_CHECK_DELAY
-    # continue
-    #####
 
     echo "[docker_leader] Checking on managers..."
 
     # check on managers
     # demote unreachable hosts if leader has lost patience
-    # error=$(sudo docker info --format '{{.Swarm.Error}}')
-    if [ -z $error ]; then # TODO: change method of detecting errors
-        managers=$(sudo docker node ls --filter "role=manager" --format "{{.Hostname}}:{{.ManagerStatus}}")
-        echo $managers
-    fi
-    for line in $managers; do
-        if [ ! -z $line ]; then
-            read hostname status <<< $(echo $line | sed 's/:/ /')
-            # check manager status
-            if [ $status == "Leader" ] || [ $hostname == $MESH_IDENTITY ] || [ $hostname == $designated_leader ]; then
-                continue
-            fi
-            if [ $status == "Unreachable" ]; then
-                # add to list or update
-                if ! grep -E "^$hostname [0]+$" $MANAGER_LIST; then # not on list yet
-                    echo "$hostname 0" >> $MANAGER_LIST
-                else # already on list
-                    # filter for hostname, only take first occurance, then get count
-                    count=$(grep -E "^$hostname [0-9]+$" $MANAGER_LIST | head -1 | awk '{print $2}')
-                    new_count=$(($count+1))
-                    if [ $new_count -ge $LEADER_PATIENCE ]; then
-                        echo "[docker_leader] Lost patience with host $hostname... ($(date +%T))" >>$LOG_OUT
-                        if sudo docker node demote $hostname && sudo docker node rm $hostname; then
-                            # successfully removed from swarm
-                            remove_host_from_manager_list $hostname
-                            echo "[docker_leader] Demoted and removed host $hostname ($(date +%T))" >>$LOG_OUT
+    managers=$(sudo docker node ls --filter "role=manager" --format "{{.Hostname}}:{{.ManagerStatus}}")
+    if [ $? ]; then
+        for line in $managers; do
+            if [ ! -z $line ]; then
+                read hostname status <<< $(echo $line | sed 's/:/ /')+
+                # check manager status
+                if [ $status == "Leader" ] || [ $hostname == $MESH_IDENTITY ] || [ $hostname == $designated_leader ]; then
+                    continue
+                fi
+                if [ $status == "Unreachable" ]; then
+                    # add to list or update
+                    if ! grep -E "^$hostname [0]+$" $MANAGER_LIST; then # not on list yet
+                        echo "$hostname 0" >> $MANAGER_LIST
+                    else # already on list
+                        # filter for hostname, only take first occurance, then get count
+                        count=$(grep -E "^$hostname [0-9]+$" $MANAGER_LIST | head -1 | awk '{print $2}')
+                        new_count=$(($count+1))
+                        if [ $new_count -ge $LEADER_PATIENCE ]; then
+                            echo "[docker_leader] Lost patience with host $hostname... ($(date +%T))" >>$LOG_OUT
+                            if sudo docker node demote $hostname && sudo docker node rm $hostname; then
+                                # successfully removed from swarm
+                                remove_host_from_manager_list $hostname
+                                echo "[docker_leader] Demoted and removed host $hostname ($(date +%T))" >>$LOG_OUT
+                            fi
+                        else
+                            sed -i "s/^$hostname $count$/$hostname $new_count/" $MANAGER_LIST # replace old count
                         fi
-                    else
-                        sed -i "s/^$hostname $count$/$hostname $new_count/" $MANAGER_LIST # replace old count
                     fi
-                fi
-            else # Reachable ("role=manager" already filters workers)
-                if [ $RESET_PATIENCE == true ]; then
-                    remove_host_from_manager_list $hostname
-                fi
-            fi # check manager status
-        fi
-    done
+                else # Reachable ("role=manager" already filters workers)
+                    if [ $RESET_PATIENCE == true ]; then
+                        remove_host_from_manager_list $hostname
+                    fi
+                fi # check manager status
+            fi
+        done
+    fi
 
     # promote new hosts if below ideal count
     # note: it would be better to modify this in the future so that an uneven manager count is avoided
     echo "[docker_leader] Checking for promotions..." >>$LOG_OUT
     current_manager_count=""
     current_manager_count=$(sudo docker node ls --filter "role=manager" --format "{{.Hostname}}" | wc -w)
-    if [ ! -z $current_manager_count ] && [ $current_manager_count -lt $IDEAL_MANAGER_COUNT ]; then
+    if [ $? ] && [ ! -z $current_manager_count ] && [ $current_manager_count -lt $IDEAL_MANAGER_COUNT ]; then
         echo "[docker_leader] Found only $current_manager_count managers though $IDEAL_MANAGER_COUNT are desired ($(date +%T))"
         node_ls=$(sudo docker node ls --filter "role=worker" \
             --filter "node.label=can_become_manager=true" \
